+++ modules/alloy/src/main/ts/ephox/alloy/behaviour/coupling/CouplingApis.ts
@@ -1,10 +1,15 @@
+import { Optional } from '@ephox/katamari';
+
 import { AlloyComponent } from '../../api/component/ComponentApi';
 import { CouplingState, CouplingConfig } from './CouplingTypes';
 
 const getCoupled = (component: AlloyComponent, coupleConfig: CouplingConfig, coupleState: CouplingState, name: string): AlloyComponent =>
-  // console.log('state', coupleState.readState());
   coupleState.getOrCreate(component, coupleConfig, name);
 
+const getExistingCoupled = (component: AlloyComponent, coupleConfig: CouplingConfig, coupleState: CouplingState, name: string): Optional<AlloyComponent> =>
+  coupleState.getExisting(component, coupleConfig, name);
+
 export {
-  getCoupled
+  getCoupled,
+  getExistingCoupled
 };