+++ modules/tinymce/src/themes/silver/main/ts/ui/menus/menu/MenuChoice.ts
@@ -24,7 +24,10 @@ export const createPartialChoiceMenu = (
   const hasIcons = MenuUtils.menuHasIcons(items);
   const presetItemTypes = presets !== 'color' ? 'normal' : 'color';
   const alloyItems = createChoiceItems(items, onItemValueHandler, columns, presetItemTypes, itemResponse, select, providersBackstage);
-  return MenuUtils.createPartialMenuWithAlloyItems(value, hasIcons, alloyItems, columns, presets);
+  const menuLayout: MenuUtils.MenuLayoutType = {
+    menuType: presets
+  };
+  return MenuUtils.createPartialMenuWithAlloyItems(value, hasIcons, alloyItems, columns, menuLayout);
 };
 
 export const createChoiceItems = (